Fix NullPointer when creating a new OrganizationUser
Auditable module would attempt to reference the created ORM by ID
(by creating a new row in `audits`), but would fail, as the ID column
was null.

This is due to the organization_user table first starting out as a Pivot.
It is no longer a Pivot, as it contains other attributes and hence
deserves an ID column.

This fixed the NullPointer.