Embed app config into HTML, do noy load via XHR
When the JS frontend app is started, it needs some config
about current app state and logged in user.

This was loaded via an API endpoint, but it causes a blocking
delay for one round-trip.

In order to avoid this and speed up the app, let's embed
the initial config to the HTML, loaded during the first
request.